+++ Api/src/main/java/games/alejandrocoria/mapfrontiers/api/model/FrontierShapeType.java
@@ -1,6 +1,0 @@
-package games.alejandrocoria.mapfrontiers.api.model;
-
-public enum FrontierShapeType {
-    VERTEX,
-    CHUNK
-}